Fixed macOS Builds

+++ ntgcalls/src/utils/hardware_info.cpp
@@ -2,10 +2,13 @@
 // Created by Laky64 on 02/03/2024.
 //
 
+#include <ntgcalls/utils/hardware_info.hpp>
+
 #if defined(IS_LINUX) || defined(IS_ANDROID)
 #include <unistd.h>
+#elif IS_MACOS
+#include <sys/resource.h>
 #endif
-#include <ntgcalls/utils/hardware_info.hpp>
 
 namespace ntgcalls {
     HardwareInfo::HardwareInfo() {
@@ -13,7 +16,7 @@ namespace ntgcalls {
         FILETIME ftime, fsys, fuser;
         SYSTEM_INFO sysInfo;
         GetSystemInfo(&sysInfo);
-        numProcessors = sysInfo.dwNumberOfProcessors;
+        numProcessors = static_cast<int>(sysInfo.dwNumberOfProcessors);
         GetSystemTimeAsFileTime(&ftime);
         memcpy(&lastCPU, &ftime, sizeof(FILETIME));
